+++ GameEngine/GameEngine/Timing.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <SDL.h>
+
 namespace GameEngine
 {
 	class FPSLimiter
@@ -19,7 +21,7 @@ namespace GameEngine
 
 		float m_fps = 0.0f;
 		float m_maxFPS = 60.0f;
-		float m_frameTime = 0.0f;
+		Uint32 m_frameTime = 0;
 		unsigned int m_startTicks = 0;
 	};
 }
