+++ GameEngine/ZombieGame/Gun.cpp
@@ -32,7 +32,7 @@ void Gun::update(bool isMouseDown, const glm::vec2& position, const glm::vec2& d
 
 void Gun::fire(const glm::vec2& position, const glm::vec2& direction, std::vector<Bullet>& bullets)
 {
-	std::mt19937 randomEngine(time(nullptr));
+	std::mt19937 randomEngine((unsigned int)time(nullptr));
 	std::uniform_real_distribution<float> randomRotate(-m_spread, m_spread);
 
 	m_fireEffect.play();
