+++ GameEngine/ZombieGame/Zombie.cpp
@@ -44,7 +44,7 @@ Human * Zombie::getNearestHuman(std::vector<Human*>& humans)
 	Human* closestHuman = nullptr;
 	float smallestDistance = 9999999.0f;
 
-	for (int i = 0; i < humans.size(); i++)
+	for (size_t i = 0; i < humans.size(); i++)
 	{
 		glm::vec2 distVec = humans[i]->getPosition() - m_position;
 		float distance = glm::length(distVec);
